Move ConfigHandlerException into a separate file
