+++ src/itemviewformaction.cpp
@@ -5,6 +5,7 @@
 #include <sstream>
 
 #include "config.h"
+#include "confighandlerexception.h"
 #include "exceptions.h"
 #include "fmtstrformatter.h"
 #include "itemrenderer.h"
