+++ src/keymap.cpp
@@ -4,7 +4,7 @@
 #include <vector>
 
 #include "config.h"
-#include "exceptions.h"
+#include "confighandlerexception.h"
 #include "logger.h"
 #include "strprintf.h"
 #include "utils.h"
