+++ test/keymap.cpp
@@ -2,7 +2,7 @@
 
 #include "3rd-party/catch.hpp"
 
-#include "exceptions.h"
+#include "confighandlerexception.h"
 
 using namespace newsboat;
