+++ test/regexmanager.cpp
@@ -1,7 +1,7 @@
 #include "regexmanager.h"
 
 #include "3rd-party/catch.hpp"
-#include "exceptions.h"
+#include "confighandlerexception.h"
 
 using namespace newsboat;
